requirements.txt
praw==7.8.1
boto3==1.34.39
botocore==1.34.39
python-dotenv==1.0.0
numpy>=1.26.4,<2
pandas==2.0.2
pyathena==3.0.0
weaviate-client==4.9.6
langchain>=0.3.27
langchain-core>=0.3.76,<1.0.0
langchain-openai>=0.2.11
langchain-anthropic>=0.2.4
langchain-text-splitters>=0.3.8
langsmith>=0.3.45
aws-xray-sdk==2.12.0
anthropic==0.49.0
opensearch-py==2.2.0
requests==2.31.0
requests-aws4auth==1.2.3
rank_bm25==0.2.1
mypy-boto3-dynamodb==1.37.0
qdrant-client==1.13.3